publish-commit-bottles: approve PRs after bottle push
We want to enable the dismissal of reviews after pushes to a PR branch,
but that means that approving reviews are dismissed after @BrewTestBot
pushes the bottle commit(s) to a PR.

To avoid PRs needing a second review before merging, let's have
@BrewTestBot approve the PR after the bottle push.

To support this, I've added a few checks for consistency of the HEAD
SHA1 in the runner's checkout and on the PR branch to avoid merging
commits we don't intend to. These checks are not airtight, but I believe
that they're better than nothing.

Finally: I've CCed myself in all comments posted after failures so I can
get a better look at making this workflow handle those situations
better.

One important change worth noting here is that the
`publish-commit-bottles` workflow _can_ be run on PRs that have not yet
been approved. These changes will mean that anyone who can trigger this
workflow can merge PRs without approval. However, it is not difficult
for maintainers who have access to this workflow to circumvent a
required review, so this is not a particularly significant drawback.